requirements.txt
# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: Apache-2.0
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# Please keep requirements sorted alphabetically
attrs==25.1.0
av==14.2.0
better-profanity==0.7.0
einops==0.7.0
einx==0.1.3
huggingface-hub==0.29.2
hydra-core==1.3.2
imageio[ffmpeg]
iopath==0.1.10
loguru==0.7.3
mediapy==1.2.2
megatron-core==0.10.0
natsort==8.4.0
nltk==3.9.1
opencv-contrib-python==4.10.0.84
peft==0.14.0
pillow==10.4.0
pycocotools
retinaface-py==0.0.2
rtmlib==0.0.13
sentencepiece==0.2.0
termcolor==2.5.0
torch==2.6.0
torchvision==0.21.0
transformers==4.48.0
